+++ .storybook/config.js
@@ -1,9 +1,10 @@
+import { configure } from '@storybook/react';
+import { storybookWindowObjects, storybookDefaultMocks, storybookWpStyles } from './../scripts/storybook';
+
 // Storybook import order is really important because it won't work in any configuration. Be careful when changing stuff here.
 
-import { configure } from '@storybook/react';
-
 // @WP Editor set default window objects.
-require('EightshiftBlocksStorybookWindowObjects');
+storybookWindowObjects();
 
 // Run all storybook stories.
 configure([
@@ -11,16 +12,14 @@ configure([
   require.context('./../blocks/init/src/blocks/custom', true, /docs\/story.js$/),
   require.context('./../blocks/init/src/blocks/wrapper', true, /docs\/story.js$/),
   require.context('./../blocks/playground', true, /story.js$/),
+  require.context('./../scripts/components', true, /docs\/story.js$/),
 ], module);
 
 // @WP Editor set default categories.
-require('EightshiftBlocksStorybookDefaultCategories');
+storybookDefaultMocks();
 
 // WP styles.
-require('EightshiftBlocksStorybookWpStyles');
-
-// @WP Editor Styles.
-require('EightshiftBlocksStorybookEditorStyles');
+storybookWpStyles();
 
 // Project styles.
 require('./../blocks/init/assets/styles/application.scss');