+++ blocks/init/src/blocks/custom/card-list/card-list-block.js
@@ -1,7 +1,7 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
 import { Fragment } from '@wordpress/element';
 import { InspectorControls } from '@wordpress/block-editor';
-import { getActions } from '@eightshift/frontend-libs';
+import { getActions } from '@eightshift/frontend-libs/scripts/editor';
 import manifest from './manifest.json';
 import { CardListEditor } from './components/card-list-editor';
 import { CardListOptions } from './components/card-list-options';