+++ blocks/init/src/blocks/custom/image/image-block.js
@@ -1,7 +1,7 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
 import { Fragment } from '@wordpress/element';
 import { InspectorControls } from '@wordpress/block-editor';
-import { getActions } from '@eightshift/frontend-libs';
+import { getActions } from '@eightshift/frontend-libs/scripts/editor';
 import manifest from './manifest.json';
 import { ImageEditor } from './components/image-editor';
 import { ImageOptions } from './components/image-options';