+++ blocks/init/storybook/webpack.config.js
@@ -3,6 +3,5 @@
 const path = require('path');
 
 module.exports = ({ config }) => {
-  const nodeModules = path.resolve(__dirname, '..');
-  return require('./../node_modules/@eightshift/frontend-libs/webpack/storybook')({ config }, nodeModules);
+  return require('./../node_modules/@eightshift/frontend-libs/webpack/storybook')({ config }, path.resolve(__dirname, '..'));
 };