+++ scripts/components/index.js
@@ -1,0 +1,2 @@
+export { ColorPaletteCustom } from './color-palette-custom/color-palette-custom';
+export { HeadingLevel } from './heading-level/heading-level';