+++ scripts/editor/index.js
@@ -1,0 +1,7 @@
+export { getActions } from './get-actions';
+export { getBlockEditComponent } from './get-block-edit-component';
+export { getPalleteColors } from './get-pallete-colors';
+export { registerBlock } from './register-block';
+export { registerBlocks } from './register-blocks';
+export { ucfirst } from './ucfirst';
+export { withWrapper } from './with-wrapper';